Removed a few redundant casts to eliminate warnings.

+++ src/net/java/dev/typecast/ot/table/GlyfCompositeDescript.java
@@ -60,7 +60,7 @@ import java.util.ArrayList;
  * or more simple glyphs, usually with some sort of transformation applied to
  * each.
  *
- * @version $Id: GlyfCompositeDescript.java,v 1.3 2007-01-24 09:47:48 davidsch Exp $
+ * @version $Id: GlyfCompositeDescript.java,v 1.4 2007-01-24 09:54:44 davidsch Exp $
  * @author <a href="mailto:davidsch@dev.java.net">David Schweinsberg</a>
  */
 public class GlyfCompositeDescript extends GlyfDescript {
@@ -123,7 +123,7 @@ public class GlyfCompositeDescript extends GlyfDescript {
             int y = gd.getYCoordinate(n);
             short x1 = (short) c.scaleX(x, y);
             x1 += c.getXTranslate();
-            return (short) x1;
+            return x1;
         }
         return 0;
     }
@@ -137,7 +137,7 @@ public class GlyfCompositeDescript extends GlyfDescript {
             int y = gd.getYCoordinate(n);
             short y1 = (short) c.scaleY(x, y);
             y1 += c.getYTranslate();
-            return (short) y1;
+            return y1;
         }
         return 0;
     }
